Added STM32U585 support and expanded DFM test analysis * Added B-U585I-IOT02A/OpenOCD support with Cortex-M33, startup code and STM32U5 dependencies * Extended the FreeRTOS suite with board selection, serial autodetection and hardware tests * Added configurable agent reviews, metrics, benchmarking and systematic fault injection * Aligned the FreeRTOS and Zephyr review tooling, tests and documentation

+++ ZephyrDemo/modules-staging/modules/debug/percepio/DFM/dfmTaskMonitor.c
@@ -32,16 +32,16 @@ static void prvOnTaskAnomaly(TraceTaskMonitorCallbackData_t *pxData)
 
 	if (pxData->uxCPULoad > pxData->uxHighLimit)
 	{
-		snprintf(cDfmPrintBuffer, sizeof(cDfmPrintBuffer), "Task %s executed more than expected (%u %%).\n", pxData->acName, pxData->uxCPULoad);
+		snprintf(cDfmPrintBuffer, sizeof(cDfmPrintBuffer), "Task %s executed more than expected (%u %%).\n", pxData->acName, (unsigned int)pxData->uxCPULoad);
 	}
 	else if (pxData->uxCPULoad < pxData->uxLowLimit)
 	{
-		snprintf(cDfmPrintBuffer, sizeof(cDfmPrintBuffer), "Task %s executed less than expected (%u %%).\n", pxData->acName, pxData->uxCPULoad);
+		snprintf(cDfmPrintBuffer, sizeof(cDfmPrintBuffer), "Task %s executed less than expected (%u %%).\n", pxData->acName, (unsigned int)pxData->uxCPULoad);
 	}
 	else
 	{
 		/* Not supposed to happen... */
-		snprintf(cDfmPrintBuffer, sizeof(cDfmPrintBuffer), "TaskMonitor alert, Task %s, CPU load %u %%, unknown reason.\n", pxData->acName, pxData->uxCPULoad);
+		snprintf(cDfmPrintBuffer, sizeof(cDfmPrintBuffer), "TaskMonitor alert, Task %s, CPU load %u %%, unknown reason.\n", pxData->acName, (unsigned int)pxData->uxCPULoad);
 	}
         
     if (xDfmAlertBegin(DFM_TYPE_TASKMONITOR_ANOMALY, cDfmPrintBuffer, &xAlertHandle) != DFM_SUCCESS)
